Breakbeatmaker: further refactoring (two)
